elastic-beats: Enable networking during do_compile
Since Kirkstone, builds are more strictly sandboxed and only
do_fetch can have network access.

There's a way to opt-out. This band-aid is used here.

In the long run, the submodules used to build beats should be
used in SRC_URI in a way like the "gosu" recipe uses it
in meta-openembedded.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>